Split test_math.py
